Add files via upload
This is the hand over hand locking scheme that was briefly mentioned in Ch. 29 in Concurrency. This locking scheme is a revision of the concurrent linked list. The way this works is that the program first locks the head (assuming the thread calls successfully), and then locks the next node, after the next node is locked successfully, the previous node is unlocked. This allows one thread to stay ahead of the other(s) (assuming two or more threads were created on the same list.) and prevents two or more threads from accessing the same node.  This enables high concurrency, with a downside of this implementation is that it has a high degree of overhead, considering each thread must lock, lock the next, and then unlock. This implementation works well, but considering the high overhead, it takes a considerable amount of time on large lists, and is slower than other implementations (BestPerformance mentioned in the book).